refactor: CouponManager의 쿠폰 생성 로직 CouponForm으로 분리
